worker/requirements.txt
playwright==1.41.0
redis==5.0.1
google-generativeai==0.3.2
